+++ tests/extension_tests/deferred_bindings_tests/test_deferred_bindings.py
@@ -3,6 +3,7 @@
 import sys
 import unittest
 
+import azure.functions as func
 from tests.utils import testutils
 
 from azure_functions_worker import protos
@@ -20,6 +21,14 @@ DEFERRED_BINDINGS_ENABLED_DIR = testutils.EXTENSION_TESTS_FOLDER / \
     'deferred_bindings_tests' / \
     'deferred_bindings_functions' / \
     'deferred_bindings_enabled'
+DEFERRED_BINDINGS_DISABLED_DIR = testutils.EXTENSION_TESTS_FOLDER / \
+    'deferred_bindings_tests' / \
+    'deferred_bindings_functions' / \
+    'deferred_bindings_disabled'
+DEFERRED_BINDINGS_ENABLED_DUAL_DIR = testutils.EXTENSION_TESTS_FOLDER / \
+    'deferred_bindings_tests' / \
+    'deferred_bindings_functions' / \
+    'deferred_bindings_enabled_dual'
 
 
 class MockMBD:
@@ -35,6 +44,7 @@ class MockMBD:
                                               "is only supported for 3.9+.")
 class TestDeferredBindingsEnabled(testutils.AsyncTestCase):
 
+    @testutils.retryable_test(3, 5)
     async def test_deferred_bindings_enabled_metadata(self):
         async with testutils.start_mockhost(
                 script_root=DEFERRED_BINDINGS_ENABLED_DIR) as host:
@@ -45,6 +55,7 @@ class TestDeferredBindingsEnabled(testutils.AsyncTestCase):
                              protos.StatusResult.Success)
         del sys.modules['function_app']
 
+    @testutils.retryable_test(3, 5)
     async def test_deferred_bindings_enabled_log(self):
         async with testutils.start_mockhost(
                 script_root=DEFERRED_BINDINGS_ENABLED_DIR) as host:
@@ -62,7 +73,69 @@ class TestDeferredBindingsEnabled(testutils.AsyncTestCase):
 
 @unittest.skipIf(sys.version_info.minor <= 8, "The base extension"
                                               "is only supported for 3.9+.")
-class TestDeferredBindingsEnabledHelpers(testutils.AsyncTestCase):
+class TestDeferredBindingsDisabled(testutils.AsyncTestCase):
+
+    @testutils.retryable_test(3, 5)
+    async def test_deferred_bindings_disabled_metadata(self):
+        async with testutils.start_mockhost(
+                script_root=DEFERRED_BINDINGS_DISABLED_DIR) as host:
+            await host.init_worker()
+            r = await host.get_functions_metadata()
+            self.assertIsInstance(r.response, protos.FunctionMetadataResponse)
+            self.assertEqual(r.response.result.status,
+                             protos.StatusResult.Success)
+        del sys.modules['function_app']
+
+    @testutils.retryable_test(3, 5)
+    async def test_deferred_bindings_disabled_log(self):
+        async with testutils.start_mockhost(
+                script_root=DEFERRED_BINDINGS_DISABLED_DIR) as host:
+            await host.init_worker()
+            r = await host.get_functions_metadata()
+            disabled_log_present = False
+            for log in r.logs:
+                message = log.message
+                if "Deferred bindings enabled: False" in message:
+                    disabled_log_present = True
+                    break
+            self.assertTrue(disabled_log_present)
+        del sys.modules['function_app']
+
+
+@unittest.skipIf(sys.version_info.minor <= 8, "The base extension"
+                                              "is only supported for 3.9+.")
+class TestDeferredBindingsEnabledDual(testutils.AsyncTestCase):
+
+    @testutils.retryable_test(3, 5)
+    async def test_deferred_bindings_dual_metadata(self):
+        async with testutils.start_mockhost(
+                script_root=DEFERRED_BINDINGS_ENABLED_DUAL_DIR) as host:
+            await host.init_worker()
+            r = await host.get_functions_metadata()
+            self.assertIsInstance(r.response, protos.FunctionMetadataResponse)
+            self.assertEqual(r.response.result.status,
+                             protos.StatusResult.Success)
+        del sys.modules['function_app']
+
+    @testutils.retryable_test(3, 5)
+    async def test_deferred_bindings_dual_enabled_log(self):
+        async with testutils.start_mockhost(
+                script_root=DEFERRED_BINDINGS_ENABLED_DUAL_DIR) as host:
+            await host.init_worker()
+            r = await host.get_functions_metadata()
+            enabled_log_present = False
+            for log in r.logs:
+                message = log.message
+                if "Deferred bindings enabled: True" in message:
+                    enabled_log_present = True
+                    break
+            self.assertTrue(enabled_log_present)
+        del sys.modules['function_app']
+
+
+@unittest.skipIf(sys.version_info.minor <= 8, "The base extension"
+                                              "is only supported for 3.9+.")
+class TestDeferredBindingsHelpers(testutils.AsyncTestCase):
 
     def test_deferred_bindings_enabled_decode(self):
         binding = BlobClientConverter
@@ -95,28 +168,31 @@ class TestDeferredBindingsEnabledHelpers(testutils.AsyncTestCase):
 
         The second represents if the current binding is deferred binding. If this is
         True, then deferred bindings must also be enabled at the function level.
-
-        Test type 1: type is supported, deferred_bindings_enabled is not yet set
-        Test type 2: type is supported, deferred_bindings_enabled is already set
         """
         async with testutils.start_mockhost(
                 script_root=DEFERRED_BINDINGS_ENABLED_DIR) as host:
             await host.init_worker()
+
+            # Type is not supported, deferred_bindings_enabled is not yet set
+            self.assertEqual(meta.check_deferred_bindings_enabled(
+                func.InputStream, False), (False, False))
+
+            # Type is not supported, deferred_bindings_enabled already set
+            self.assertEqual(meta.check_deferred_bindings_enabled(
+                func.InputStream, True), (True, False))
+
+            # Type is supported, deferred_bindings_enabled is not yet set
             self.assertEqual(meta.check_deferred_bindings_enabled(
                 BlobClient, False), (True, True))
-
-            self.assertEqual(meta.check_deferred_bindings_enabled(
-                BlobClient, True), (True, True))
-
             self.assertEqual(meta.check_deferred_bindings_enabled(
                 ContainerClient, False), (True, True))
-
-            self.assertEqual(meta.check_deferred_bindings_enabled(
-                ContainerClient, True), (True, True))
-
             self.assertEqual(meta.check_deferred_bindings_enabled(
                 StorageStreamDownloader, False), (True, True))
 
+            # Type is supported, deferred_bindings_enabled is already set
+            self.assertEqual(meta.check_deferred_bindings_enabled(
+                BlobClient, True), (True, True))
+            self.assertEqual(meta.check_deferred_bindings_enabled(
+                ContainerClient, True), (True, True))
             self.assertEqual(meta.check_deferred_bindings_enabled(
                 StorageStreamDownloader, True), (True, True))
-        del sys.modules['function_app']
